Move stat under lib/
