+++ src/lib/stat.js
